requirements.txt
Pillow
RPi.GPIO
opencv-python
email
picamera
pyyaml
smtplib
ssl
